upload image

+++ app/src/main/java/uz/koinot/stadion/data/model/CreateStadium.kt
@@ -10,8 +10,8 @@ class CreateStadium(
     var opening_time: String,
     var closing_time: String,
     var change_price_time: String,
-    var price_day_time: Int,
-    var price_night_time: Int,
+    var price_day_time: Double,
+    var price_night_time: Double,
     var active: Boolean,
     var width: Int,
     var height: Int
